Add support viewing and downloading the original files
This basically reads the resource from the local disk. I think this
is no security hole because a file needs to exist in the index and on
disk to be retrievable. Still, let me use this commit message as
another point to advice you to make neither ElasticSearch nor the
search page itself accessible from the outside. A local search engine
should only accept local queries.