Add the initial idea of a Dockerfile
This is mainly to give the search app a quick spin. The Docker image
needs a separate ElasticSearch instance (in for example the Docker
"elasticsearch" container) to store the data. The directories to scan
should be mounted into the container using the -v: switch, but that
and the rescanning are not yet implemented.